Refactoring/clarifying, moving HiGlassTabView to own file.
